refactor: centralize mention notifications
